fix(codeql): widen integer products before casting to double
CodeQL raised three error-level cs/loss-of-precision alerts on this branch, all the same shape: an
integer multiplication whose result is then cast to double, so the product is formed in the narrow
domain and only the result widens.

Worth recording that none of the three can actually overflow. MaxCompressedDim is 128, so 2 * c is at
most 256; and totalFrames is already a long, so totalFrames * _melChannels would need on the order of
1e17 frames before it mattered. CodeQL also did not flag the strictly larger (2 * d) * (2 * c) two
lines above one of them, which is a further sign the rule fires on the cast site rather than on real
risk.

The widening is applied anyway: it is behaviour-identical for every reachable value, it is the
canonical form for this rule, and it clears an error-level gate that is currently failing the check.

+++ src/Audio/TextToSpeech/AlignedTextToMelModelBase.cs
@@ -239,8 +239,10 @@ public abstract partial class AlignedTextToMelModelBase<T> : AudioNeuralNetworkB
             if (includeOutput) outputs.Add(new OutputBucket(group.Value.ToArray(), predictedMel, predictedLogs));
         }
 
-        melLoss = Engine.TensorMultiplyScalar(melLoss, NumOps.FromDouble(1.0 / (totalFrames * _melChannels)));
-        priorLoss = Engine.TensorMultiplyScalar(priorLoss, NumOps.FromDouble(0.5 / (totalFrames * _melChannels)));
+        // Widen before multiplying: the product is formed in the integer domain otherwise, and only the
+        // result is cast, so the division would inherit whatever that product truncated to.
+        melLoss = Engine.TensorMultiplyScalar(melLoss, NumOps.FromDouble(1.0 / ((double)totalFrames * _melChannels)));
+        priorLoss = Engine.TensorMultiplyScalar(priorLoss, NumOps.FromDouble(0.5 / ((double)totalFrames * _melChannels)));
         durationLoss = Engine.TensorMultiplyScalar(durationLoss, NumOps.FromDouble(1.0 / totalTokens));
         var total = Engine.TensorAdd(Engine.TensorAdd(melLoss, priorLoss), durationLoss);
         var output = includeOutput ? AssembleOutput(outputs, durations, batch.MelLengths.ToArray(), tokens.Shape[1]) : null;
